Move tests outside of the script
